requirements.txt
streamlit
pandas
plotly
numpy
torch 
torchvision
ultralytics
Pillow